Remove casting to int for strtol()

+++ proj2.c
@@ -143,7 +143,7 @@ int parse_p(int argc, char ** argv, params_t * pars) {
 
 	// argv[1] = NO, argv[2] = NH, argv[3] = TI, argv[4] = TB
 	for (int i = 1; i < argc; i++) {
-		int tmp = (int)strtol(argv[i], &endPtr, 10);
+		int tmp = strtol(argv[i], &endPtr, 10);
 
 		if (tmp < 0 || *endPtr != 0) {
 			fprintf(stderr, "Parsed value is not a positive integer.\n");
@@ -151,10 +151,10 @@ int parse_p(int argc, char ** argv, params_t * pars) {
 		}
 	}
 
-	pars->n_oxygens = (int)strtol(argv[1], &endPtr, 10);
-	pars->n_hydrogens = (int)strtol(argv[2], &endPtr, 10);
-	pars->t_i = (int)strtol(argv[3], &endPtr, 10);
-	pars->t_b = (int)strtol(argv[4], &endPtr, 10);
+	pars->n_oxygens = strtol(argv[1], &endPtr, 10);
+	pars->n_hydrogens = strtol(argv[2], &endPtr, 10);
+	pars->t_i = strtol(argv[3], &endPtr, 10);
+	pars->t_b = strtol(argv[4], &endPtr, 10);
 
 	if (pars->t_b > MAX_TB || pars->t_i > MAX_TI) {
 		fprintf(stderr, "TB or TI have to be in range <0,1000> incl.\n");
